WIP nested comments now render on create, but not on initial load
